Fix: 반응형 수정

+++ src/pages/EditPage.jsx
@@ -48,8 +48,12 @@ function EditPage() {
   const bgImg = bgData.backgroundImageURL;
 
   const handleDeleteRecipients = async () => {
-    await deleteRecipientDAsync(id);
-    navigate('/list');
+    const userConfirmed = window.confirm('정말로 삭제하시겠습니까?');
+
+    if (userConfirmed) {
+      await deleteRecipientDAsync(id);
+      navigate('/list');
+    }
   };
 
   return (
